Cleaned up and noted code Fixed IE display
Commented throughout the document on what the classes do, Added hover effects for the links however now there is a white space above the header that i can not seem to get rid of.  Also Need to find a fix for the bottom banner as well works perfect in full screen but when the screen size changes it causes issues.